+++ src/common/WordsScrambleGame/enums/directions.enum.ts
@@ -1,0 +1,10 @@
+export enum Directions {
+  Right = 'right',
+  Left = 'left',
+  Down = 'down',
+  Up = 'up',
+  Diagonal = 'diagonal',
+  Vertical = 'vertical',
+  Horizontal = 'horizontal',
+  Default = 'default',
+}